runtime: add block parameters
parameters provide a mechanism so that changes to a value of a block can
be centralized and made threadsafe.  Parameters can be extended so that
the registered values can be updated via tags, RPC, etc without extra
code in each block.  This is only implemented here on the multiply_const
block as a proof of concept