runtime: Define runtime_deps as dependencies in dependency_declare
This is required to have compile definitionsa and link flags from
dependencies used for builds including this dependency.

This issue was found through spdlog which has a compile definition
to switch between bundled and external fmtlib and this definition was
available for runtime lib but not for other libraries depending on
runtime lib.

Signed-off-by: Andrej Rode <mail@andrejro.de>